# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sipelayar API",
    "version": "1.0.0",
    "description": "API documentation for Sipelayar application",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "V1 Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token issued on login"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key issued to clients"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/balances": {
      "get": {
        "summary": "Get all balances",
        "tags": [
          "Balances"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "minBalance",
            "schema": {
              "type": "number",
              "format": "float"
            },
            "description": "Minimum balance filter"
          },
          {
            "in": "query",
            "name": "maxBalance",
            "schema": {
              "type": "number",
              "format": "float"
            },
            "description": "Maximum balance filter"
          }
        ],
        "responses": {
          "200": {
            "description": "List of balances"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/balances/device/{deviceId}": {
      "get": {
        "summary": "Get balance for a specific device",
        "tags": [
          "Balances"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "deviceId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Device ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Device balance details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Device not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/balances/check": {
      "post": {
        "summary": "Check device balance by device key",
        "tags": [
          "Balances"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "deviceKey"
                ],
                "properties": {
                  "deviceKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device balance checked successfully"
          },
          "400": {
            "description": "Device key is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Device is inactive"
          },
          "404": {
            "description": "Device not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/balances/update": {
      "post": {
        "summary": "Update balance for a device",
        "tags": [
          "Balances"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "deviceId",
                  "amount"
                ],
                "properties": {
                  "deviceId": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number",
                    "format": "float",
                    "description": "Positive or negative amount to update"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Balance updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Device not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/balances/{deviceId}": {
      "patch": {
        "summary": "Partially update device balance",
        "tags": [
          "Balances"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "deviceId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Device ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "format": "float",
                    "description": "Amount to add (positive) or subtract (negative)"
                  },
                  "lastToken": {
                    "type": "string",
                    "description": "Last token used (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Balance partially updated successfully"
          },
          "400": {
            "description": "Validation error or no fields to update"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - no permission to update this balance"
          },
          "404": {
            "description": "Device or balance not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/balances/stats": {
      "get": {
        "summary": "Get balance statistics",
        "tags": [
          "Balances"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Balance statistics"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}